Completed the finanace and activity module basics

+++ Logger.DAL/Core/Repository.cs
@@ -16,7 +16,7 @@ namespace Logger.DAL.Core
             Context = context;
         }
 
-        public TEntity Get(int id)
+        public TEntity Get(long id)
         {
             // Here we are working with a DbContext, not PlutoContext. So we don't have DbSets 
             // such as Courses or Authors, and we need to use the generic Set() method to access them.
